# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Daily Shop Ledger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entries:
    get:
      operationId: listEntries
      tags: [entries]
      summary: List all entries (optionally filtered)
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [cash_in, cash_out]
        - name: payment_method
          in: query
          schema:
            type: string
            enum: [cash, digital]
        - name: date
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
        - name: end_date
          in: query
          schema:
            type: string
        - name: deleted
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createEntry
      tags: [entries]
      summary: Create a new ledger entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntryBody"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /entries/{id}:
    get:
      operationId: getEntry
      tags: [entries]
      summary: Get a single entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateEntry
      tags: [entries]
      summary: Update an entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntryBody"
      responses:
        "200":
          description: Entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteEntry
      tags: [entries]
      summary: Soft-delete an entry (move to recycle bin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /entries/{id}/restore:
    patch:
      operationId: restoreEntry
      tags: [entries]
      summary: Restore entry from recycle bin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /entries/{id}/profit:
    patch:
      operationId: updateEntryProfit
      tags: [entries]
      summary: Set or update profit for an entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfitBody"
      responses:
        "200":
          description: Profit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reports/entries:
    get:
      operationId: getEntriesReport
      tags: [reports]
      summary: Get entries report for a period
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Entries report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntriesReport"
  /reports/profit:
    get:
      operationId: getProfitReport
      tags: [reports]
      summary: Get profit report for a period
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Profit report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitReport"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Entry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [cash_in, cash_out]
        amount:
          type: number
        description:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [cash, digital]
        profit:
          type: ["number", "null"]
        isCredit:
          type: boolean
        customerName:
          type: ["string", "null"]
        deletedAt:
          type: ["string", "null"]
        entryDate:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - type
        - amount
        - paymentMethod
        - isCredit
        - entryDate
        - createdAt
        - updatedAt
    CreateEntryBody:
      type: object
      properties:
        type:
          type: string
          enum: [cash_in, cash_out]
        amount:
          type: number
        description:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [cash, digital]
        isCredit:
          type: boolean
        customerName:
          type: ["string", "null"]
        entryDate:
          type: ["string", "null"]
      required:
        - type
        - amount
        - paymentMethod
    UpdateEntryBody:
      type: object
      properties:
        type:
          type: string
          enum: [cash_in, cash_out]
        amount:
          type: number
        description:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [cash, digital]
        isCredit:
          type: boolean
        customerName:
          type: ["string", "null"]
        entryDate:
          type: ["string", "null"]
    UpdateProfitBody:
      type: object
      properties:
        profit:
          type: ["number", "null"]
      required:
        - profit
...
    EntriesReport:
      type: object
      properties:
        period:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        totalCashIn:
          type: number
        totalCashOut:
          type: number
        cashBalance:
          type: number
        digitalBalance:
          type: number
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
      required:
        - period
        - startDate
        - endDate
        - totalCashIn
        - totalCashOut
        - cashBalance
        - digitalBalance
        - entries
    ProfitReport:
      type: object
      properties:
        period:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        totalProfit:
          type: number
        entriesWithProfit:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
      required:
        - period
        - startDate
        - endDate
        - totalProfit
        - entriesWithProfit
...