# lib/api-zod/src/generated/types/profitReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Daily Shop Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { Entry } from "./entry";

export interface ProfitReport {
  period: string;
  startDate: string;
  endDate: string;
  totalProfit: number;
  entriesWithProfit: Entry[];
}